refactor: eliminate all re-exports and barrel index files
Remove barrel index files (components/index.ts, plugins/transformers/index.ts,
plugins/emitters/index.ts, plugins/filters/index.ts) and update all consumers
to import directly from source modules. Also remove individual re-exports from
escape.ts, search.ts, and linkfavicons.ts.

Add no-re-export rule to CLAUDE.md Code Style section and Lessons Learned.
