--- quartz/components/index.ts
@@ -1,25 +1,0 @@
-import ArticleTitle from "./ArticleTitle"
-import AuthorList from "./Authors"
-import { Backlinks } from "./Backlinks"
-import ContentMeta from "./ContentMeta"
-import Footer from "./Footer"
-import Head from "./Head"
-import Navbar from "./Navbar"
-import NotFound from "./pages/404"
-import Content from "./pages/Content"
-import TagContent from "./pages/TagContent"
-import TableOfContents from "./TableOfContents"
-
-export {
-  ArticleTitle,
-  Content,
-  TagContent,
-  Head,
-  ContentMeta,
-  TableOfContents,
-  Navbar,
-  Backlinks,
-  Footer,
-  NotFound,
-  AuthorList,
-}
