+++ quartz/components/scripts/search.ts
@@ -1,9 +1,9 @@
+import escapeRegExp from "escape-string-regexp"
 import FlexSearch, { type ContextOptions } from "flexsearch"
 
 import { type ContentDetails } from "../../plugins/emitters/contentIndex"
 import { replaceEmojiConvertArrows } from "../../plugins/transformers/twemoji"
 import { tabletBreakpoint, mobileBreakpoint } from "../../styles/variables"
-import { escapeRegExp } from "../../util/escape"
 import { type FullSlug, resolveRelative } from "../../util/path"
 import { simpleConstants } from "../constants"
 import { registerEscapeHandler, removeAllChildren, debounce } from "./component_script_utils"
@@ -173,9 +173,6 @@ export function match(searchTerm: string, text: string, trim?: boolean) {
   }
   return `${beginning}${slice}${end}`
 }
-
-// Re-export escapeRegExp from centralized escape utilities
-export { escapeRegExp } from "../../util/escape"
 
 /**
  * Creates a span element with the class "match" and the given text
