+++ quartz/plugins/emitters/404.tsx
@@ -1,7 +1,8 @@
 import { sharedPageComponents } from "../../../config/quartz/quartz.layout"
 import { type FullPageLayout } from "../../cfg"
-import { ContentMeta, NotFound } from "../../components"
 import BodyConstructor from "../../components/Body"
+import ContentMeta from "../../components/ContentMeta"
+import NotFound from "../../components/pages/404"
 import { pageResources, renderPage } from "../../components/renderPage"
 import { type QuartzComponentProps } from "../../components/types"
 import DepGraph from "../../depgraph"
