--- quartz/plugins/emitters/index.ts
@@ -1,12 +1,0 @@
-export { ContentPage } from "./contentPage"
-export { TagPage } from "./tagPage"
-export { ContentIndex } from "./contentIndex"
-export { AliasRedirects } from "./aliases"
-export { Assets } from "./assets"
-export { Static } from "./static"
-export { ComponentResources } from "./componentResources"
-export { NotFoundPage } from "./404"
-export { CNAME } from "./cname"
-export { AllTagsPage } from "./allTagsPage"
-export { AllPostsPage as RecentPostsPage } from "./allPostsPage"
-export { PopulateContainers } from "./populateContainers"
