--- quartz/plugins/filters/index.ts
@@ -1,1 +1,0 @@
-export { RemoveDrafts } from "./draft"
